requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.0
click==8.1.7
colorama==0.4.6
flash==1.0.3
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
scikit-learn==1.4.2
scipy==1.13.0
threadpoolctl==3.5.0
Werkzeug==3.0.3
WTForms==3.1.2
